+++ tasks/populate-email-template-css.js
@@ -1,16 +1,23 @@
 const fs = require("fs");
 
-const emailNunjucksLocation = "package/nationalarchives/templates/email.njk";
+const replacementString = "<!-- COMPILED_EMAIL_CSS -->";
 
-fs.readFile(emailNunjucksLocation, "utf8", (err, data) => {
-  if (err) {
-    console.error(err);
-    return;
-  }
-  fs.readFile(
-    "package/nationalarchives/email.css",
-    "utf8",
-    (err, compiledCSS) => {
+fs.readFile(
+  "package/nationalarchives/email.css",
+  "utf8",
+  (err, compiledCSS) => {
+    if (err) {
+      console.error(err);
+      return;
+    }
+
+    compiledCSS = compiledCSS
+      .replace("/*# sourceMappingURL=email.css.map */", "")
+      .trim();
+
+    const emailNunjucksLocation =
+      "package/nationalarchives/templates/layouts/email.njk";
+    fs.readFile(emailNunjucksLocation, "utf8", (err, data) => {
       if (err) {
         console.error(err);
         return;
@@ -18,8 +25,8 @@ fs.readFile(emailNunjucksLocation, "utf8", (err, data) => {
       fs.writeFile(
         emailNunjucksLocation,
         data.replace(
-          "<!-- COMPILED_CSS -->",
-          `<style media="all" type="text/css">${compiledCSS.replace("/*# sourceMappingURL=email.css.map */", "").trim()}</style>`,
+          replacementString,
+          `<style media="all" type="text/css">${compiledCSS}</style>`,
         ),
         (err) => {
           if (err) {
@@ -27,6 +34,27 @@ fs.readFile(emailNunjucksLocation, "utf8", (err, data) => {
           }
         },
       );
-    },
-  );
-});
+    });
+
+    const templateFixturesLocation =
+      "package/nationalarchives/templates/fixtures.json";
+    fs.readFile(templateFixturesLocation, "utf8", (err, data) => {
+      if (err) {
+        console.error(err);
+        return;
+      }
+      fs.writeFile(
+        templateFixturesLocation,
+        data.replace(
+          replacementString,
+          `<style media=\\"all\\" type=\\"text/css\\">${compiledCSS.replace(/\\/g, "\\\\").replace(/"/g, '\\"')}</style>`,
+        ),
+        (err) => {
+          if (err) {
+            return console.error(err);
+          }
+        },
+      );
+    });
+  },
+);